requirements.txt
torch>=2.2
torchvision>=0.17
tqdm>=4.66
pyyaml>=6.0
Pillow>=10.0
